dao almost finished

+++ clockin/src/main/java/dao/HabitDao.java
@@ -1,6 +1,7 @@
 package dao;
 
 import model.Habit;
+import model.Record;
 import model.User;
 import utils.DatabaseUtil;
 import utils.DateUtil;
@@ -51,17 +52,90 @@ public class HabitDao {
         return flag;
     }
 
-    public boolean deleteById(int id) {
+    public boolean deleteById(int id, String userId) {
         //删除一条习惯
         boolean flag = true;
         Connection connection = null;
         PreparedStatement pst = null;
-        String sql = "delete from Habit where id = ?";
+        String sql = "delete from Habit where id = ? and user_id = ?";
         try {
             connection = DatabaseUtil.getConnection();
             connection.setAutoCommit(false);
             pst = connection.prepareStatement(sql);
             pst.setInt(1, id);
+            pst.setString(2, userId);
+            if(pst.executeUpdate() != 1) {
+                flag = false;
+            }
+            connection.commit();
+            connection.setAutoCommit(true);
+        } catch (SQLException e) {
+            flag = false;
+            try {
+                connection.rollback();
+            } catch (SQLException e1) {
+                e1.printStackTrace();
+            }
+            e.printStackTrace();
+        } finally {
+            DatabaseUtil.close(null, pst, connection);
+        }
+        return flag;
+    }
+
+    public boolean updateHabit(Habit habit) {
+        boolean flag = true;
+        Connection connection = null;
+        PreparedStatement pst = null;
+        String sql = "update Habit set name = ?, icon = ?, color = ?, category = ?, weekday = ?, flag_auto = ? where id = ? and user_id = ?";
+        try {
+            connection = DatabaseUtil.getConnection();
+            connection.setAutoCommit(false);
+            pst = connection.prepareStatement(sql);
+            pst.setString(1, habit.getName());
+            pst.setInt(2, habit.getIcon());
+            pst.setInt(3, habit.getColor());
+            pst.setInt(4, habit.getCategory());
+            pst.setInt(5, habit.getWeekday());
+            pst.setBoolean(6, habit.isFlag_auto());
+            pst.setInt(7, habit.getId());
+            pst.setString(8, habit.getUserId());
+            if(pst.executeUpdate() != 1) {
+                flag = false;
+            }
+            connection.commit();
+            connection.setAutoCommit(true);
+        } catch (SQLException e) {
+            flag = false;
+            try {
+                connection.rollback();
+            } catch (SQLException e1) {
+                e1.printStackTrace();
+            }
+            e.printStackTrace();
+        } finally {
+            DatabaseUtil.close(null, pst, connection);
+        }
+        return flag;
+    }
+
+    public boolean clockIn(int id, boolean op, String userId) {
+        //用来打卡,op=false为取消打卡，op为true为打卡
+        boolean flag = true;
+        Connection connection = null;
+        PreparedStatement pst = null;
+        String sql;
+        if(op) {
+            sql = "update Habit set clockin_days = clockin_days + 1 where id = ? and user_id = ?";
+        } else {
+            sql = "update Habit set clockin_days = clockin_days - 1 where id = ? and user_id = ?";
+        }
+        try {
+            connection = DatabaseUtil.getConnection();
+            connection.setAutoCommit(false);
+            pst = connection.prepareStatement(sql);
+            pst.setInt(1, id);
+            pst.setString(2, userId);
             if(pst.executeUpdate() != 1) {
                 flag = false;
             }
@@ -82,44 +156,92 @@ public class HabitDao {
     }
 
     public static void main(String[] args) {
-        /*HabitDao a = new HabitDao();
-        Habit h = new Habit();
-        h.setUserId("User");
+        HabitDao a = new HabitDao();
+        //AddHabit测试
+        /*Habit h = new Habit();
+        h.setUserId("user");
         h.setName("name");
         h.setIcon(1);
         h.setColor(2);
         h.setCategory(3);
         h.setWeekday(4);
         h.setFlag_auto(true);
+        a.addHabit(h);
+        h.setUserId("user2");
+        h.setName("name2");
         a.addHabit(h);*/
-        System.out.println(DateUtil.getNowTime());
-        System.out.println(DateUtil.getNowDateTime("yyyyMMdd"));
+
+        //deleteById测试
+        /*boolean flag = a.deleteById(5, "user");
+        System.out.println(flag);*/
+
+        //updateHabit测试
+        /*Habit n = new Habit();
+        n.setUserId("user");
+        n.setId(7);
+        n.setName("k");
+        n.setIcon(9);
+        n.setColor(9);
+        n.setCategory(9);
+        n.setFlag_auto(false);
+        n.setWeekday(9);
+        boolean flag = a.updateHabit(n);
+        System.out.println(flag);*/
+
+        //clockin测试
+        /*boolean flag = a.clockIn(7, true, "user2");
+        System.out.println(flag);*/
+
+        //findAllHabitByUserId测试
+        /*List<Habit> list = a.findAllHabitByUserId("user2", 3);
+        for (Habit habit:list
+            ) {
+            System.out.println(habit.getId());
+        }*/
+
+        //findById测试
+        /*Habit habit = a.findById(6, "user2");
+        System.out.println(habit);*/
+
+        //findTodayHabit
+        List<Habit> list = a.findTodayHabit("user2");
+
+//        System.out.println(DateUtil.getDayOfWeek());
+        //System.out.println(DateUtil.getNowTime());
+        //System.out.println(DateUtil.getNowDateTime("yyyyMMdd"));
     }
-    /*
-    public List<Comment> findCommentByPicId(int phoId) {
+
+    public List<Habit> findAllHabitByUserId(String userId, int category) {
+        //查询所有习惯时使用，-1代表所有习惯
         Connection connection = null;
         PreparedStatement pst = null;
         ResultSet rs = null;
-        List<Comment> list = new ArrayList<>();
-        String sql = "select * from comment where phoid = ?";
+        List<Habit> list = new ArrayList<>();
+        String sql = "select * from Habit where user_id = ? and category = ? order by create_time asc";
+        if(category == -1) {
+            sql = "select * from Habit where user_id = ? order by create_time asc";
+        }
         try {
             connection = DatabaseUtil.getConnection();
             connection.setAutoCommit(false);
             pst = connection.prepareStatement(sql);
-            pst.setInt(1, phoId);
+            pst.setString(1, userId);
+            if(category != -1) {
+                pst.setInt(2, category);
+            }
             rs = pst.executeQuery();
-            Comment comment;
+            Habit habit;
             while(rs.next()) {
-                comment = new Comment();
-                comment.setId(rs.getInt("id"));
-                UserDaoImpl impl = new UserDaoImpl();
-                User user = impl.findById(rs.getInt("userid"));
-                comment.setUsername(user.getName());
-                comment.setContent(rs.getString("comment"));
-                comment.setDate(rs.getString("date"));
-                comment.setPicid(rs.getInt("phoid"));
-                comment.setUserid(rs.getInt("userid"));
-                list.add(comment);
+                habit = new Habit();
+                habit.setId(rs.getInt("id"));
+                habit.setName(rs.getString("name"));
+                habit.setIcon(rs.getInt("icon"));
+                habit.setColor(rs.getInt("color"));
+                habit.setClockinDays(rs.getInt("clockin_days"));
+                habit.setCreateTime(rs.getString("create_time"));
+                habit.setWeekday(rs.getInt("weekday"));
+                habit.setCategory(rs.getInt("category"));
+                list.add(habit);
             }
             connection.commit();
             connection.setAutoCommit(true);
@@ -132,14 +254,15 @@ public class HabitDao {
             e.printStackTrace();
         }
         return list;
-    }*/
+    }
 
-    public Habit findById(int id) {
+    public Habit findById(int id, String userId) {
+        //查详情时使用该方法，需要判断是否为空
         Habit habit = new Habit();
         Connection connection = null;
         PreparedStatement pst = null;
         ResultSet rs = null;
-        String sql = "select * from Habit where id = ?";
+        String sql = "select * from Habit where id = ? and user_id = ?";
         connection = DatabaseUtil.getConnection();
         try {
             connection.setAutoCommit(false);
@@ -149,6 +272,7 @@ public class HabitDao {
             else {
                 pst = connection.prepareStatement(sql);
                 pst.setInt(1, id);
+                pst.setString(2, userId);
                 rs = pst.executeQuery();
                 if(rs.next()) {
                     habit.setId(rs.getInt("id"));
@@ -160,7 +284,11 @@ public class HabitDao {
                     habit.setCreateTime(rs.getString("create_time"));
                     habit.setClockinDays(rs.getInt("clockin_days"));
                     habit.setFlag_auto(rs.getBoolean("flag_auto"));
-    ????????????????????????????????????????
+                    RecordDao dao = new RecordDao();
+                    int days = dao.getInsistDays(habit.getId(), userId);
+                    habit.setInsistDays(days);
+                    boolean today = dao.isTodayClockIn(id, userId);
+                    habit.setFlag_today(today);
                 }
                 else {
                     habit = null;
@@ -180,5 +308,96 @@ public class HabitDao {
             DatabaseUtil.close(rs, pst, connection);
         }
         return habit;
+    }
+
+    public List<Habit> findTodayHabit(String userId) {
+        Connection connection = null;
+        PreparedStatement pst = null;
+        ResultSet rs = null;
+        List<Habit> list = new ArrayList<>();
+        String sql = "select * from Habit where user_id = ? order by category asc, create_time asc";
+        int dayOfWeek = DateUtil.getDayOfWeek();
+        int temp = 1 << (7 - dayOfWeek);
+        //System.out.println(temp);
+        try {
+            connection = DatabaseUtil.getConnection();
+            connection.setAutoCommit(false);
+            pst = connection.prepareStatement(sql);
+            pst.setString(1, userId);
+            rs = pst.executeQuery();
+            Habit habit;
+            while(rs.next()) {
+                habit = new Habit();
+                habit.setWeekday(rs.getInt("weekday"));
+//                System.out.print(habit.getWeekday() + "    ");
+//                System.out.println(habit.getWeekday() & temp);
+                if((habit.getWeekday() & temp) == temp) {
+                    habit.setId(rs.getInt("id"));
+                    habit.setName(rs.getString("name"));
+                    habit.setIcon(rs.getInt("icon"));
+                    habit.setColor(rs.getInt("color"));
+                    habit.setClockinDays(rs.getInt("clockin_days"));
+                    habit.setCreateTime(rs.getString("create_time"));
+                    habit.setCategory(rs.getInt("category"));
+                    habit.setFlag_auto(rs.getBoolean("flag_auto"));
+                    RecordDao dao = new RecordDao();
+                    boolean flagToday = dao.isTodayClockIn(habit.getId(), userId);
+                    int days = dao.getInsistDays(habit.getId(), userId);
+                    habit.setFlag_today(flagToday);
+                    habit.setInsistDays(days);
+                    list.add(habit);
+                }
+            }
+            connection.commit();
+            connection.setAutoCommit(true);
+        } catch (SQLException e) {
+            try {
+                connection.rollback();
+            } catch (SQLException e1) {
+                e1.printStackTrace();
+            }
+            e.printStackTrace();
+        }
+        return list;
+    }
+
+    public boolean habitPermission(int habitId, String userId) {
+        Connection connection = null;
+        PreparedStatement pst = null;
+        ResultSet rs = null;
+        String sql = "select * from Habit where id = ? and user_id = ?";
+        boolean flag = true;
+        connection = DatabaseUtil.getConnection();
+        try {
+            connection.setAutoCommit(false);
+            if(habitId <= 0) {
+                flag = false;
+            }
+            else {
+                pst = connection.prepareStatement(sql);
+                pst.setInt(1, habitId);
+                pst.setString(2, userId);
+                rs = pst.executeQuery();
+                if(rs.next()) {
+                    flag = true;
+                }
+                else {
+                    flag = false;
+                }
+            }
+            connection.commit();
+            connection.setAutoCommit(true);
+        } catch (SQLException e) {
+            flag = false;
+            try {
+                connection.rollback();
+            } catch (SQLException e1) {
+                e1.printStackTrace();
+            }
+            e.printStackTrace();
+        } finally {
+            DatabaseUtil.close(rs, pst, connection);
+        }
+        return flag;
     }
 }
